% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{colorize}
\alias{colorize}
\title{Colorize according to a column in a data frame}
\usage{
colorize(df, col.att, color, new.col.name)
}
\arguments{
\item{df}{a data frame}

\item{col.att}{an integer indicating the column with which to create the gradient}

\item{color}{a character vector of length two indicating the starting point of the gradient and the ending point of the gradiant}

\item{new.col.name}{a string indicating the gradient column name added to argument df}
}
\value{
a data frame.
}
\description{
Add a new column to a data frame of gradient color based on a specific column of the data frame
}
\details{
argument df with an additional column corresponding to the gradient column based on argument att.
}
\keyword{internal}
