% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Data frame}
\format{
An object of class \code{data.frame} with 20 rows and 6 columns.
}
\usage{
df
}
\description{
This is a simulated data frame of individual characteristics
}
\author{
Sebastian Sosa
}
\keyword{internal}
