% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{data}
\name{m}
\alias{m}
\title{Matrix data frame}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 20 columns.
}
\usage{
m
}
\description{
This is a simulated data frame of individual characteristics
}
\author{
Sebastian Sosa
}
\keyword{internal}
