% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mat.to.edgl}
\alias{mat.to.edgl}
\title{Matrix to edge list}
\usage{
mat.to.edgl(M, sym = FALSE, erase.diag = TRUE)
}
\arguments{
\item{M}{a square adjacency matrix.}

\item{sym}{if \emph{TRUE}, it extracts the lower triangle of the matrix only.}

\item{erase.diag}{if \emph{TRUE}, it omits diagonals.}
}
\value{
a data frame representing an edge list.
}
\description{
Converts a square adjacency matrix into a data frame with three columns representing an edge list. Columns are: actor, receiver and weight.
}
\details{
This function convert a matrix into an edgelist.
}
\author{
Sebastian Sosa
}
\keyword{internal}
