% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{shape}
\alias{shape}
\title{Specify shapes according to a column in a data frame}
\usage{
shape(vec, char)
}
\arguments{
\item{vec}{a  vector}

\item{char}{a character vector}
}
\value{
a numeric vector.
}
\description{
Specify shapes according to a column in a data frame
}
\details{
Converts a character vector into a numeric vector.
}
\keyword{internal}
