% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{vis.net.format.att}
\alias{vis.net.format.att}
\title{specifies shapes according to a column in a data frame}
\usage{
vis.net.format.att(
  df,
  col.id = NULL,
  col.size = NULL,
  color = NULL,
  col.color = NULL,
  shapes = NULL,
  col.shape = NULL,
  strokeCol = NULL,
  col.strokeCol = NULL,
  col.stroke = NULL,
  node.opacity = NULL
)
}
\arguments{
\item{df}{a data frame with node information.}

\item{col.id}{a character or numeric vector indicating the column name or index of node Ids.}

\item{col.size}{a character or numeric vector indicating the column name or index of node size.}

\item{color}{a vector character of length two indicating the range of color for node color.}

\item{col.color}{a character or numeric vector indicating the column name or index of node color.}

\item{shapes}{a character vector indicating the type of shape per category. 7 shapes are available: "circle", "cross", "diamond", "rectangle, "star", "triangle", "y"}

\item{col.shape}{a character or numeric vector indicating the column name or index of node shape.}

\item{strokeCol}{a vector character of length two indicating the range of color for node stroke color.}

\item{col.strokeCol}{a character or numeric vector indicating the column name or index of node stroke color.}

\item{col.stroke}{a character or numeric vector indicating the column name or index of node stroke.}

\item{node.opacity}{a character or numeric vector indicating the column name or index of node opacity.}
}
\value{
a data frame.
}
\description{
specifies shapes according to a column in a data frame
}
\details{
format a data frame with informations relative to vis.net function.
}
\keyword{internal}
